run error

+++ keras/datasets/cifar10.py
@@ -13,7 +13,7 @@ def load_data(test_split=0.1, seed=113):
     path = get_file(dirname, origin=origin, untar=True)
 
     nb_samples = 50000
-    X = np.zeros((nb_samples, 3, 32, 32), dtype="uint8")
+    X = np.zeros((nb_samples, 3, 32, 32), dtype="float64")
     y = np.zeros((nb_samples,), dtype="uint8")
     for i in range(1, 6):
         fpath = path + '/data_batch_' + str(i)
